Dired config updates
Omit .. in dired when subdirectories have been inserted.
(Usually .. would be omitted anyway, but I have custom settings to keep
it around by default, so this commit is to override that in circumstances
where it is no longer desirable.)

Reinstated the old `dired-load-hook' code, as that approach seems to
work after all. I'm unsure what my original issue was, although I do
note that these autoloaded functions are still reporting themselves as
autoloaded after the library really has loaded.  I suspect this was
confusing me at the time (and seems like a bug).

Some sandbox testing suggests that this ought to be a reliable config
for using dired-x features.