Trigger SQL upcase code with `post-command-hook' rather than self-insert
I realised that `post-self-insert-hook' will never catch RET, as it
doesn't self-insert.

Also stop looking at the input events at all. In general, the events
bound to the relevant commands are not hard-coded, so that's not a
sensible approach.

Lastly, added basic error handling. Without any, the function could
be removed automatically from the hook. I've seen this happen, but
am not currently sure whether there are bugs in the current code, or
if it only happened while I was breaking things in development.

TODO: I believe I should actually use `after-change-functions'.