Change `my-keys-minor-mode' to a globalized (rather than global) mode
The minibuffer inhibiting code has taken to disabling the global mode
globally, instead of buffer-locally.  It would be good to establish
what has changed in Emacs to cause this change of behaviour, but for
now changing to a globalized minor mode (with the minibuffer logic
within a `my-keys-turn-on' function) circumvents the issue.