Add `my-copy-region-as-kill' bound to "M-w"
By default this just calls `copy-region-as-kill'; but with a prefix arg it
instead calls `my-copy-region-unindented'.

In the latter case, C-u unindents the text completely, whilst a numeric arg
causes the copied text to be indented by (only) PAD columns (i.e. after being
fully unindented, it is then re-indented by that amount).