Add `my-kill-region' bound to "C-w"
By default this just calls `kill-region'; but with a prefix arg it
instead calls `my-copy-region-unindented' for copying the region to
the kill ring (and then `delete-region' to delete it).

In the latter case, C-u unindents the text completely, whilst a
numeric arg causes the copied text to be indented by (only) PAD
columns (i.e. after being fully unindented, it is then re-indented
by that amount).