psysh.el: Ensure psysh uses interactive mode
Version 0.10 of PsySH introduced a non-interactive mode which is
selected automatically when there is no terminal, so we must enable
'interactiveMode' explicitly, as we do with 'colorMode'.

Earlier versions of PsySH will just ignore this config, so the change
is backwards-compatible; however we cannot use the PHP constant
\Psy\Configuration::INTERACTIVE_MODE_FORCED if it isn't defined, so
we instead use the value 'forced' and, for consistency, make the same
change to our existing use of \Psy\Configuration::COLOR_MODE_FORCED.

These constant values are defined in psysh's src/Configuration.php file.