psysh.el: Naming consistency
Rename `psysh-beginning-of-line-or-indentation' to
`psysh-move-beginning-of-line-or-indentation', for consistency with
`psysh-move-beginning-of-line'.

This is still a bit awkward, as behind the scenes we are calling
`beginning-of-line' rather than `move-beginning-of-line' (in both
commands), so there's an argument for *neither* of the psysh commands
having 'move-' in their name; however `psysh-move-beginning-of-line'
is *used* as a remapping for `move-beginning-of-line' (C-a), so it
makes sense from that perspective.

I'm unsure whether the differences between `beginning-of-line' and
`move-beginning-of-line' are relevant to psysh at all.