Tweaking `drupal-tags-autoupdate-callback' for remote files
This still doesn't check whether `tags-file-name' and
`default-directory' are on the same host, but it's an improvement on
the previous behaviour (which was using tramp paths in the shell
commands).

Now redirecting 'stat' output to dev null, so we no longer see when
it's trying to stat something unexpected.  If I still have dumb bugs,
then this is going to obscure them.