Remove sass-mode and its dependency haml-mode
I've had some problems with sass-mode, and then found that the in-built
scss-mode seemed to work fine, so I'm deferring to that unless I encounter
some reason to switch back.

My history for SASS support seems a little confused...

I initially added scss-mode in commit dc2848a5d248671cf564d6bcad8b216a006229fc
and that library may or may not be the ancestor of the code which is currently
a part of the standard css-mode.el; but I added that in 2015 and the in-built
mode was added in 2014.  Mostly likely I was using an older version of Emacs
at that time.

Only a year ago I replaced my external scss-mode.el with sass-mode.el and its
haml-mode.el dependency, and removed scss-mode.el; however I suspect that I
was unaware that the external library was shadowing another scss-mode.  This
was commit b4b814564821c455a323060adc6d62f2fe50ef9c (adding the new libraries)
and commit f43e25827fc1579b9dfe56a9e54a9632643ba7de (removing the old one).

Hopefully the standard scss-mode proves to be sufficient.