Hide trailing whitespace in all special modes
This eliminates a heap bunch of explicit or indirect derivatives.

The exception being `sx-cmopose-mode' which may have been in that list
by mistake, as it's a `markdown-mode' derivative and so I'm not sure
I'd really want to hide trailing whitespace.  I suspect that may have
been on account of empty lines in indented code blocks being a common
thing; but as nowadays the triple-backtick syntax for code formatting
is preferred on stackexchange, I think I can omit this (especially as
I've not been using sx.el anytime recently).