Allow backups to be made much more frequently
By default, only the very first save of a file during a session may
create a backup of the previous version.  When sessions can last for
weeks or months, and buffers are kept for the duration, this default
is painfully inadequate!

Somehow no one has ever added a user option for customizing this,
so I should probably submit a patch for this feature...