Keep/show trailing whitespace when editing the diary
Whitespace is significant to the format, and determines what will be
displayed in the fancy diary view.  In particular, indented continuation
lines for a diary entry would be broken by any genuinely empty line, so
we need such "empty" lines to actually contain a (trailing) space, and
therefore we need to be able to see trailing whitespace.

(info "(emacs)Format of Diary File") explains:

> Each entry in the diary file describes one event and consists of one
> or more lines.  An entry always begins with a date specification at the
> left margin.  The rest of the entry is simply text to describe the
> event.  If the entry has more than one line, then the lines after the
> first must begin with whitespace to indicate they continue a previous
> entry.  Lines that do not begin with valid dates and do not continue a
> preceding entry are ignored.  Here’s an example:

n.b. This feature of the format can be used for commenting entries, and
to control what is displayed in appt.el reminders (omitting excessive
details if desired).