28.1: Remove broken/unused etags-stack.el library
I never used this, it no longer works, and I don't see an updated
version, so can't check whether it was useful.  I'm assuming that
I don't need it.

n.b. It looks like I sourced it from:
http://www.emacswiki.org/emacs/etags-stack.el
but that is the same non-working version I'm removing now, and
nor is there an ELPA package in the regular archives.