Deal with relative url() file paths in SASS 'partials/*' files
These are relative to the parent directory, and so this allows us to
visit them from the url() text.

It means any other `find-file' usage is starting from the wrong
directory, but the use cases are constrained and I think in practice
I tend to use `dired-jump' to see what the other "partials" SASS files
are named in order to visit any of them, so this works out ok.