Stop requiring external libraries at init time
I can't even remember whether I knew that :features were `require'd
originally, but if so then I'd certainly forgotten.  I absolutely
don't need all of this to be happening at init time, and especially
not for the libraries that I'm not habitually using (and some of these
could most likely be removed entirely).

If it turns out to be annoying for any of these to not be pre-loaded,
I can always revert those cases.