Silence byte-compiler warning
Eliminate this:

> In my-keys-have-priority: my-keybindings.el:428:8:
> Warning: value from call to `assq-delete-all' is unused

It's a false-positive, because the call to `assq-delete-all' only ever
happened in the situation where the warning doesn't apply.

I could have used `warning-suppress-log-types' to suppress the
`byte-compile--log-warning-for-byte-compile' call to `display-warning'
(with type `bytecomp'); or I could have added a redundant setq; but
this refactoring is better than either of those options.